Minor Changes
I fixed a few things, but there are still a few problems. Characters are
not immediately being removed from the stack after they die. They still
are attacking until the next round when the stack gets reset. Also, the
navigation is off. I was able to go out of the array's bounds. The map
is supposed to be 10 x 10, so that might be easier to check than trying
to check it with any size map we give it.