i failed to do the added text
so the map loading occurs before the loading image gets clicked.

i tried to put a waiting loop in to delay loading the maze, i failed.

i tried to load the map from the gamescreen, but the code directly after
needed the map.

i tried to add another layer to do it, i failed i am at my wits end and
need help with this.